[ADD] hr_payroll_allowance
Adds the concept of payroll allowances
The aim of this addon is to take away the complexity associated with setting
up salary rules and editing service structures to reflect allowances being paid
to employees.
Also from experience a lot of companies in West Africa (may also apply to other
regions) do not have standardized pay structures with employees in the same job
having different allowances.