[10.0][MIG] port hr_holidays_legal_leave for 10.0 (#262)
[MIG] port hr_holidays_legal_leave for 10.0

hr.config.settings is not inherited anymore, because one is not existed
in 10.0 and there is no anything similar. Only new menu can be added,
but I don't like this solution. I have added computes fields to
hr.holidays.status instead

res.company is not inherited anymore, because there is no HR section in
Company