[10.0] hr_public_holidays: allow to change next years days to be created
When creating next year's public holidays using the wizard, holidays being flagged
with variable=True (Date may change) were created using last year's date although
this date will not be the same in 99% of the cases.
This commit adds a new model with an onchange to allow the user to set next years
holiday date in the wizard before the creation of related holidays.