[ADD] hr_employee_calendar_planning: Employee Calendar Planning
This module allows to manage employee working time with profiles by date
intervals.

The profiles are regular working time calendars, but they are treated as
master ones here, allowing you to compose complexes working times by dates.

Under the hook, a unique working time is created for each employee with the
proper composition for not affecting the rest of the functionality linked to
this model.

Installation
============

During the installation of the module, current working times are split by
start/end dates for having consistent data, and the potential new composed
calendar planning is saved instead on the employee.

Configuration
=============

1. Go to *Employees > Employees*.
2. Open or create a new one.
3. On the "Public Information" page, fill the section "Calendar planning" with:

   * Starting date (optional).
   * Ending date (optional).
   * Working time to apply during that date interval.

Known issues / Roadmap
======================

* Add a wizard for generating next year calendar planning based on current one
  in batch.
* Add constraint for avoiding planning lines overlapping.
* Avoid the regeneration of whole private calendars each time a change is
  detected.