[ADD] hr_attendance_sheet
[IMP] LINT & Cleanup

Reduce warnings due to similar field labels
Code formatting cleanup for LINT warnings

[FIX] Decoration on wrong view

Also removed unused api

[IMP] Removed Split Attendance

Removed split attendance so it can be put into it's own module as suggested by reviewers.

[IMP] Remove split attendance from readme

[FIX] Remove unused timedelta, time

[IMP] Updated cpmoute_duration

 Updated compute_duration to only apply lunch on first attenance and added the auto lunch field to lists.

[IMP] prevent locked sheet linkage

[IMP] delta code cleanup

[IMP] added auto lunch override

The managers should be able to override the auto lunch on an attendance. This provides that capability.

[FIX] Bi-Weekly sheet creation

Sheets were being created for both weekly and bi-weekly option was set. Removing the 'or' out of the statement seemed to resolve the issue.

[IMP] Remove attendance inline edit

[FIX] sheet creation not using company properly

Also improved the existing sheet lookup

[IMP] Added multi-company record rule

[IMP] Auto lunch adjustment

Added logic to auto lunch so it won't apply the auto lunch if they take the proper lunch time between attendances.

[IMP] Updated settings verbiage

[IMP] Added a 'Attendance Admin' for approval

This adds another layer of approval by adding an 'Attendance Admin' field to the departments. This way a person (not manager) that is put in charge of managing attendances for a department can now administer attendances.

[IMP] Attendance Sheet Revision

[IMP] Company Date Check

[IMP] Updates and Cleanup

[IMP] Black and Tests

[IMP] Increased Tests Coverage

[IMP] LINT and Tests

[IMP] More Test Coverage

[IMP] more tests coverage

[IMP] Tests

[IMP] Test Coverage

[MIG] Add Test Coverage

[MIG] Test Coverage

[MIG] Test Labels

[IMP] Test Coverage

[FIX] License to AGPL

[IMP] Added another test