[FIX] hr_period: Error in onchange is fixed.
Hr.period inherit from date.range model from server-tools, and, in
date_range is added an onchange to assign a possible parent when update
dates, type or company. But a error is returned when try to assign the
search result because try to assign:

date.range = hr.period

To avoid this, now the field parent_id is assigned in hr_period, and
with this avoids the error.